refactor: folder architecture
